Creating mundane-path for Path and FileName
