+++ mundane-io/src/test/scala/com/ambiata/mundane/io/DirectoriesSpec.scala
@@ -2,6 +2,7 @@ package com.ambiata.mundane.io
 
 import com.ambiata.mundane.io.Arbitraries._
 import com.ambiata.mundane.testing.RIOMatcher._
+import com.ambiata.mundane.path._
 
 import org.specs2._
 import org.scalacheck._
