+++ mundane-io/src/test/scala/com/ambiata/mundane/io/FilesSpec.scala
@@ -2,6 +2,8 @@ package com.ambiata.mundane.io
 
 import com.ambiata.mundane.io.Arbitraries._
 import com.ambiata.mundane.testing.RIOMatcher._
+import com.ambiata.mundane.path._
+
 import org.specs2._
 import scalaz._, Scalaz._
