+++ mundane-io/src/test/scala/com/ambiata/mundane/io/LineCountSpec.scala
@@ -5,6 +5,7 @@ import com.ambiata.mundane.control._
 import com.ambiata.mundane.data.Lists
 import com.ambiata.mundane.io.Arbitraries._
 import com.ambiata.mundane.testing.RIOMatcher._
+import com.ambiata.mundane.path._
 
 import java.io._
