+++ mundane-path/src/test/scala/com/ambiata/path/Arbitraries.scala
@@ -1,4 +1,4 @@
-package com.ambiata.mundane.io
+package com.ambiata.mundane.path
 
 import org.scalacheck._
 import Arbitrary._
